modified publish_gh_pages script

+++ publish-gh-pages.sh
@@ -7,15 +7,19 @@ python make_zip.py
 rm -rf _site
 git clone . _site
 
-cd _site
+mkdir -p _site
+if cd _site; then
+  git checkout gh-pages
+  git pull
 
-git checkout gh-pages
-git pull
-
-rm -rf scripts node_modules assets
-unzip -o -q ../app.zip
-git add *
-git commit -a -m "update gh-pages"
-git push
-cd ../
-rm -rf _site
+  rm -rf scripts node_modules assets
+  unzip -o -q ../app.zip
+  git add *
+  git commit -a -m "update gh-pages"
+  git push
+  cd ../
+  rm -rf _site
+else
+  echo "Failed to create _site dir"
+  exit 1
+fi;
